### 2024-06-03 #6786 @Kevin-Mattheus-Moerman
@willros yes I'm looking for a handling editor. Our bioinformatics editors have all been preoccupied with other submissions so none have been available. I'll assign one as soon as one is.
