### 2024-06-05 #6786 @csoneson
👋🏻 @chrisamiller, @KatyBrown, @kdm9 - would you be interested in reviewing this submission for JOSS?

> Fraggler: A Python Package and CLI Tool for Automated Fragment Analysis
https://github.com/openjournals/joss-reviews/issues/6786

Thanks in advance!
