### 2024-06-10 5ccdcc6
rewrite of fraggler
