+++ fraggler_rewrite/fraggler.py
@@ -1,0 +1,1305 @@
+# TODO: add rere.py as testing
+# Make a script and print the results in the terminal and capture it!
+
+import numpy as np
+from sklearn.preprocessing import PolynomialFeatures
+from sklearn.linear_model import LinearRegression
+from sklearn.pipeline import make_pipeline
+from sklearn.preprocessing import SplineTransformer
+from sklearn.metrics import mean_squared_error, r2_score
+from scipy.interpolate import UnivariateSpline
+from pathlib import Path
+from scipy import sparse
+from scipy.sparse import linalg
+from numpy.linalg import norm
+from Bio import SeqIO
+from scipy import signal
+import matplotlib.pyplot as plt
+import pandas as pd
+import altair as alt
+from lmfit.models import VoigtModel, GaussianModel, LorentzianModel
+import re
+import pandstat
+import sys
+from datetime import datetime
+import argparse
+import warnings
+import platform
+import panel as pn
+
+
+
+warnings.filterwarnings("ignore")
+# for windows user
+if platform.system() == "Windows":
+    matplotlib.use("agg")
+
+
+from ladders import LADDERS
+
+### UTILITY FUNCTIONS ###
+def baseline_arPLS(y, ratio=0.99, lam=100, niter=1000, full_output=False):
+    """
+    Taken from:
+    https://stackoverflow.com/questions/29156532/python-baseline-correction-library
+
+    from this paper:
+    https://pubs.rsc.org/en/content/articlelanding/2015/AN/C4AN01061B#!divAbstract
+    """
+    L = len(y)
+
+    diag = np.ones(L - 2)
+    D = sparse.spdiags([diag, -2 * diag, diag], [0, -1, -2], L, L - 2)
+
+    H = lam * D.dot(D.T)  # The transposes are flipped w.r.t the Algorithm on pg. 252
+
+    w = np.ones(L)
+    W = sparse.spdiags(w, 0, L, L)
+
+    crit = 1
+    count = 0
+
+    while crit > ratio:
+        z = linalg.spsolve(W + H, W * y)
+        d = y - z
+        dn = d[d < 0]
+
+        m = np.mean(dn)
+        s = np.std(dn)
+
+        w_new = 1 / (1 + np.exp(2 * (d - (2 * s - m)) / s))
+
+        crit = norm(w_new - w) / norm(w)
+
+        w = w_new
+        W.setdiag(w)  # Do not create a new matrix, just update diagonal values
+
+        count += 1
+
+        if count > niter:
+            print("Maximum number of iterations exceeded")
+            break
+
+    if full_output:
+        info = {"num_iter": count, "stop_criterion": crit}
+        return z, d, info
+    else:
+        return z
+
+
+def make_dir(outpath: str) -> None:
+    outpath = Path(outpath)
+    if not outpath.exists():
+        outpath.mkdir(parents=True)
+
+
+def get_files(in_path: str) -> list[Path]:
+    # If in_path is a directory, get a list of all .fsa files in it
+    if Path(in_path).is_dir():
+        files = [x for x in Path(in_path).iterdir() if x.suffix == ".fsa"]
+    else:
+        files = [Path(in_path)]
+    return files
+
+
+def file_exists(file):
+    if not Path(file).exists():
+        print_fail(f"{file} does not exist!")
+        sys.exit(1)
+
+
+def folder_exists(folder):
+    if Path(folder).exists():
+        print_fail(f"{folder} already exist!")
+        sys.exit(1)
+
+
+class bcolors:
+    OKBLUE = "\033[94m"
+    OKCYAN = "\033[96m"
+    OKGREEN = "\033[92m"
[…]
