+++ fraggler_rewrite/fraggler.py
@@ -1,6 +1,35 @@
 # TODO: add rere.py as testing
 # Make a script and print the results in the terminal and capture it!
 
+
+class bcolors:
+    OKBLUE = "\033[94m"
+    OKCYAN = "\033[96m"
+    OKGREEN = "\033[92m"
+    WARNING = "\033[93m"
+    FAIL = "\033[91m"
+    ENDC = "\033[0m"
+    UNDERLINE = "\033[4m"
+
+
+def print_green(text):
+    print(f"{bcolors.OKGREEN}[INFO]: {text}{bcolors.ENDC}")
+
+
+def print_warning(text):
+    print(f"{bcolors.WARNING}{bcolors.UNDERLINE}[WARNING]: {text}{bcolors.ENDC}")
+
+
+def print_fail(text):
+    print(f"{bcolors.FAIL}{bcolors.UNDERLINE}[ERROR]: {text}{bcolors.ENDC}")
+
+
+def print_blue(text):
+    print(f"{bcolors.OKBLUE}[SUMMARIZE]: {text}{bcolors.ENDC}")
+
+
+print_green(f"Starting fraggler, importing libraries...")
+
 import numpy as np
 from sklearn.preprocessing import PolynomialFeatures
 from sklearn.linear_model import LinearRegression
@@ -26,7 +55,6 @@ import argparse
 import warnings
 import platform
 import panel as pn
-
 
 
 warnings.filterwarnings("ignore")
@@ -112,32 +140,6 @@ def folder_exists(folder):
     if Path(folder).exists():
         print_fail(f"{folder} already exist!")
         sys.exit(1)
-
-
-class bcolors:
-    OKBLUE = "\033[94m"
-    OKCYAN = "\033[96m"
-    OKGREEN = "\033[92m"
-    WARNING = "\033[93m"
-    FAIL = "\033[91m"
-    ENDC = "\033[0m"
-    UNDERLINE = "\033[4m"
-
-
-def print_green(text):
-    print(f"{bcolors.OKGREEN}[INFO]: {text}{bcolors.ENDC}")
-
-
-def print_warning(text):
-    print(f"{bcolors.WARNING}{bcolors.UNDERLINE}[WARNING]: {text}{bcolors.ENDC}")
-
-
-def print_fail(text):
-    print(f"{bcolors.FAIL}{bcolors.UNDERLINE}[ERROR]: {text}{bcolors.ENDC}")
-
-
-def print_blue(text):
-    print(f"{bcolors.OKBLUE}[SUMMARIZE]: {text}{bcolors.ENDC}")
 
 
 ASCII_ART = f"""{bcolors.OKBLUE}
@@ -1036,12 +1038,13 @@ def parse_fsa(
     except:
         print_fail(f"Cannot parse {fsa}")
         sys.exit(1)
-        
-        
+
+
 ### REPORT ###
 pn.extension("tabulator")
 pn.extension("vega", sizing_mode="stretch_width", template="fast")
 pn.widgets.Tabulator.theme = "modern"
+
 
 def header(
     text: str,
@@ -1163,10 +1166,9 @@ def generate_area_report(fsa):
         text="## Peaks Table", bg_color="#04c273", height=80, textalign="left"
     )
 
-    df = (
-        fsa.identified_sample_data_peaks
-        [["basepairs", "assay", "peak_name", "model", "r_value", "quotient"]]
-    )
+    df = fsa.identified_sample_data_peaks[
+        ["basepairs", "assay", "peak_name", "model", "r_value", "quotient"]
+    ]
 
     # DataFrame Tabulator
     peaks_df_tab = pn.widgets.Tabulator(
@@ -1204,6 +1206,40 @@ def generate_area_report(fsa):
     return report
 
 
+def generate_no_peaks_report(fsa):
[…]
