+++ fraggler_rewrite/fraggler.py
@@ -298,6 +298,9 @@ def return_maxium_allowed_distance_between_size_standard_peaks(fsa, multiplier=2
 
 
 def generate_combinations(fsa):
+    """
+    depth-first search
+    """
     a = fsa.size_standard_peaks
     length = fsa.n_ladder_peaks
     distance = fsa.maxium_allowed_distance_between_size_standard_peaks
@@ -1085,6 +1088,113 @@ def make_header(name: str, date: str) -> pn.pane.Markdown:
     )
 
 
+def generate_peak_report(fsa):
+    ### ----- Raw Data ----- ###
+    channel_header = header(
+        text="## Plot of channels",
+        bg_color="#04c273",
+        height=80,
+        textalign="left",
+    )
+    # PLOT
+    channel_tab = pn.Tabs()
+    for plot, name in plot_fsa_data(fsa):
+        pane = pn.pane.Vega(plot.interactive(), sizing_mode="stretch_both", name=name)
+        channel_tab.append(pane)
+
+    channels_section = pn.Column(channel_header, channel_tab)
+
+    ### ----- Peaks ----- ###
+    peaks_header = header(
+        text="## Plot of Peaks",
+        bg_color="#04c273",
+        height=80,
+        textalign="left",
+    )
+
+    # PLOT
+    peaks_plot = plot_all_found_peaks(fsa)
+    peaks_pane = pn.pane.Matplotlib(peaks_plot, name="Peaks")
+
+    # Section
+    peaks_tab = pn.Tabs(
+        peaks_pane,
+    )
+    peaks_section = pn.Column(peaks_header, peaks_tab)
+
+    ### ----- Ladder Information ----- ###
+    ladder_header = header(
+        text="## Information about the ladder",
+        bg_color="#04c273",
+        height=80,
+        textalign="left",
+    )
+    # Ladder peak plot
+    ladder_plot = plot_size_standard_peaks(fsa)
+    ladder_peak_plot = pn.pane.Matplotlib(
+        ladder_plot,
+        name="Ladder Peak Plot",
+    )
+    # Ladder Correlation
+    model_fit = plot_model_fit(fsa)
+    ladder_correlation_plot = pn.pane.Matplotlib(
+        model_fit,
+        name="Ladder Correlation Plot",
+    )
+
+    # Section
+    ladder_tab = pn.Tabs(
+        ladder_peak_plot,
+        ladder_correlation_plot,
+    )
+    ladder_section = pn.Column(ladder_header, ladder_tab)
+
+    ### ----- Peaks dataframe ----- ###
+    dataframe_header = header(
+        text="## Peaks Table", bg_color="#04c273", height=80, textalign="left"
+    )
+    # Create dataframe
+    df = (
+        fsa.identified_sample_data_peaks
+        .assign(file_name=fsa.file_name)
+        [["basepairs", "assay", "peak_name", "file_name"]]
+    )
+    # DataFrame Tabulator
+    peaks_df_tab = pn.widgets.Tabulator(
+        df,
+        layout="fit_columns",
+        pagination="local",
+        page_size=15,
+        show_index=False,
+        name="Peaks Table",
+    )
+
+    # Section
+    dataframe_tab = pn.Tabs(peaks_df_tab)
+    dataframe_section = pn.Column(dataframe_header, dataframe_tab)
+
+    ### CREATE REPORT ###
+
+    file_name = fsa.file_name
+    date = fsa.fsa["RUND1"]
+    head = make_header(file_name, date)
+
+    all_tabs = pn.Tabs(
+        ("Channels", channels_section),
+        ("Peaks", peaks_section),
+        ("Ladder", ladder_section),
[…]
