+++ fraggler_rewrite/fraggler2/fraggler.py
@@ -1,0 +1,1515 @@
+# TODO: add rere.py as testing
+# Make a script and print the results in the terminal and capture it!
+
+
+class bcolors:
+    OKBLUE = "\033[94m"
+    OKCYAN = "\033[96m"
+    OKGREEN = "\033[92m"
+    WARNING = "\033[93m"
+    FAIL = "\033[91m"
+    ENDC = "\033[0m"
+    UNDERLINE = "\033[4m"
+
+
+def print_green(text):
+    print(f"{bcolors.OKGREEN}[INFO]: {text}{bcolors.ENDC}")
+
+
+def print_warning(text):
+    print(f"{bcolors.WARNING}{bcolors.UNDERLINE}[WARNING]: {text}{bcolors.ENDC}")
+
+
+def print_fail(text):
+    print(f"{bcolors.FAIL}{bcolors.UNDERLINE}[ERROR]: {text}{bcolors.ENDC}")
+
+
+def print_blue(text):
+    print(f"{bcolors.OKBLUE}[SUMMARIZE]: {text}{bcolors.ENDC}")
+
+
+print_green(f"Starting fraggler, importing libraries...")
+
+import numpy as np
+from sklearn.preprocessing import PolynomialFeatures
+from sklearn.linear_model import LinearRegression
+from sklearn.pipeline import make_pipeline
+from sklearn.preprocessing import SplineTransformer
+from sklearn.metrics import mean_squared_error, r2_score
+from scipy.interpolate import UnivariateSpline
+from pathlib import Path
+from scipy import sparse
+from scipy.sparse import linalg
+from numpy.linalg import norm
+from Bio import SeqIO
+from scipy import signal
+import matplotlib.pyplot as plt
+import pandas as pd
+import altair as alt
+from lmfit.models import VoigtModel, GaussianModel, LorentzianModel
+import re
+import sys
+from datetime import datetime
+import argparse
+import warnings
+import platform
+import panel as pn
+import pandas_flavor as pf
+
+
+warnings.filterwarnings("ignore")
+# for windows user
+if platform.system() == "Windows":
+    matplotlib.use("agg")
+
+
+from .ladders import LADDERS
+
+### UTILITY FUNCTIONS ###
+@pf.register_dataframe_method
+def pivot_wider(
+    df_: pd.DataFrame, index: list, names_from: list, values_from: list
+) -> pd.DataFrame:
+    df = df_.pivot(index=index, columns=names_from, values=values_from).reset_index()
+    names = [[str(y) for y in x] for x in df.columns]
+    names = ["_".join(x).strip("_") for x in names]
+    df.columns = names
+
+    return df
+
+
+def baseline_arPLS(y, ratio=0.99, lam=100, niter=1000, full_output=False):
+    """
+    Taken from:
+    https://stackoverflow.com/questions/29156532/python-baseline-correction-library
+
+    from this paper:
+    https://pubs.rsc.org/en/content/articlelanding/2015/AN/C4AN01061B#!divAbstract
+    """
+    L = len(y)
+
+    diag = np.ones(L - 2)
+    D = sparse.spdiags([diag, -2 * diag, diag], [0, -1, -2], L, L - 2)
+
+    H = lam * D.dot(D.T)  # The transposes are flipped w.r.t the Algorithm on pg. 252
+
+    w = np.ones(L)
+    W = sparse.spdiags(w, 0, L, L)
+
+    crit = 1
+    count = 0
+
+    while crit > ratio:
+        z = linalg.spsolve(W + H, W * y)
+        d = y - z
+        dn = d[d < 0]
+
+        m = np.mean(dn)
+        s = np.std(dn)
+
[…]
