### 2024-06-11 #6869 @editorialbot
[REVIEW]: Fraggler: A Python Package and CLI Tool for Automated Fragment Analysis
**Submitting author:** @willros (<a href="http://orcid.org/0000-0003-2274-7343">William Rosenbaum</a>)
**Repository:** https://github.com/Clinical-Genomics-Umea/fraggler
**Branch with paper.md** (empty if default branch):
**Version:** v3.0.3
