+++ fraggler_rewrite/fraggler2/fraggler.py
@@ -184,8 +184,8 @@ class FsaFile:
         self.file_name = self.file.parts[-1]
 
         if ladder not in LADDERS.keys():
-            raise LadderNotFoundError(f"'{ladder}' is not a valid ladder")
-            exit(1)
+            print_fail(f"'{ladder}' is not a valid ladder")
+            sys.exit(1)
         self.ladder = ladder
         self.fsa = SeqIO.read(file, "abi").annotations["abif_raw"]
         self.sample_channel = sample_channel
@@ -361,17 +361,17 @@ def calculate_best_combination_of_size_standard_peaks(fsa):
     return fsa
 
 
-def fit_size_standard_to_ladder(fsa) -> tuple:
+def fit_size_standard_to_ladder(fsa):
     """
     Returns the FsaFile with updated model and datafram with sample data
     and fitted baisepairs.
     Increase the knots until every basepair is unique.
     """
     best_combination = fsa.best_size_standard
-    n_knots = 4
+    n_knots = 3
     for _ in range(20):
         model = make_pipeline(
-            SplineTransformer(degree=3, n_knots=n_knots, extrapolation="continue"),
+            SplineTransformer(degree=2, n_knots=n_knots, extrapolation="continue"),
             LinearRegression(fit_intercept=True),
         )
 
@@ -398,23 +398,27 @@ def fit_size_standard_to_ladder(fsa) -> tuple:
             return fsa
         else:
             n_knots += 1
+            
+        # if no model could be fit
+        fsa.fitted_to_model = False
+        return fsa
 
 
 ### PLOTTING ###
 def plot_areas(fsa):
     def plot_helper(identified_peaks, df):
+        peaks = [df.loc[lambda x: x.peak_name == p] for p in df.peak_name.unique()]
         fig_areas, axs = plt.subplots(
-            1, len(identified_peaks), sharey=True, figsize=(20, 10)
+            1, len(peaks), sharey=True, figsize=(20, 10)
         )
         # if there is only one peak
-        if len(identified_peaks) == 1:
+        if len(peaks) == 1:
             axs.plot(df.basepairs, df.peaks, "o")
             axs.plot(df.basepairs, df.fitted)
             axs.set_title(f"Peak 1 area: {df['amplitude'].iloc[0]: .1f}")
             axs.grid()
         # if more than one peak
         else:
-            peaks = [df.loc[lambda x: x.peak_name == p] for p in df.peak_name.unique()]
             for i, ax in enumerate(axs):
                 ax.plot(
                     peaks[i].basepairs,
@@ -613,9 +617,8 @@ def find_peaks_agnostic(
     )
 
     if identified_peaks.shape[0] == 0:
-        # TODO: REWRITE
-        print("No peaks could be found")
-        exit(1)
+        fsa.found_peaks = "error"
+        return fsa
 
     fsa.sample_data_peaks_raw = peaks_dataframe
     fsa.identified_sample_data_peaks = identified_peaks
@@ -633,9 +636,8 @@ def read_custom_peaks(custom_peaks):
         else None
     )
     if not isinstance(custom_peaks, pd.DataFrame):
-        # TODO: CHANGE
-        print("No custom peaks could be read")
-        exit(1)
+        print_fail("No custom peaks could be read")
+        sys.exit(1)
 
     return custom_peaks
 
@@ -742,7 +744,7 @@ def find_peaks_customized(
                         .drop(columns=["distance"])
                     )
             else:
-                print("[ERROR]: column `which` must be `FIRST` or `LARGEST`")
+                print_fail("Column `which` must be `FIRST` or `LARGEST`")
                 exit(1)
 
         customized_peaks.append(df)
[…]
