### 2024-06-12 9a12cc1
large rewrite and restruture
