### 2024-06-12 e044f34
paper and tests
