+++ .github/workflows/rere_test.yaml
@@ -17,10 +17,10 @@ jobs:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v3
-    - name: Set up Python 3.11
+    - name: Set up Python 3.10
       uses: actions/setup-python@v3
       with:
-        python-version: "3.11"
+        python-version: "3.10"
     - name: Test with pytest
       run: |
         pip install .
