### 2024-06-12 b929075
tests and setup
