### 2024-06-13 97f79af
added helper functions, versions and added to tutorial
