+++ fraggler/fraggler.py
@@ -61,6 +61,7 @@ if platform.system() == "Windows":
 
 from .ladders import LADDERS
 
+
 ### UTILITY FUNCTIONS ###
 @pf.register_dataframe_method
 def pivot_wider(
@@ -86,7 +87,7 @@ def pivot_wider(
         The reshaped DataFrame with columns named according to the specified
         identifiers and value variables.
     """
-    
+
     df = df_.pivot(index=index, columns=names_from, values=values_from).reset_index()
     names = [[str(y) for y in x] for x in df.columns]
     names = ["_".join(x).strip("_") for x in names]
@@ -261,7 +262,7 @@ class FsaFile:
         find_size_standard_channel(degree=2):
             Identifies the channel with the best fit for the size standard.
         """
-        
+
         self.file = Path(file)
         self.file_name = self.file.parts[-1]
 
@@ -367,13 +368,14 @@ class FsaFile:
         Fitted Area Peaks: {self.fitted_area_peaks}
         """
 
+
 def find_size_standard_peaks(fsa):
     """
     Identifies and extracts peaks from the size standard channel of an FSA file.
 
-    This function detects peaks in the size standard channel of the given FSA file 
-    using specified height and distance criteria. The identified peaks are then 
-    sorted and limited to the maximum number of allowed peaks, which are stored 
+    This function detects peaks in the size standard channel of the given FSA file
+    using specified height and distance criteria. The identified peaks are then
+    sorted and limited to the maximum number of allowed peaks, which are stored
     in the `size_standard_peaks` attribute of the FSA file object.
 
     Parameters:
@@ -384,10 +386,10 @@ def find_size_standard_peaks(fsa):
     Returns:
     --------
     FsaFile
-        The FsaFile instance with the `size_standard_peaks` attribute updated 
+        The FsaFile instance with the `size_standard_peaks` attribute updated
         to include the identified peaks.
     """
-    
+
     found_peaks = signal.find_peaks(
         fsa.size_standard,
         height=fsa.min_size_standard_height,
@@ -455,9 +457,9 @@ def calculate_best_combination_of_size_standard_peaks(fsa):
     """
     Calculates the best combination of size standard peaks for fitting the ladder model.
 
-    This function evaluates different combinations of size standard peaks to determine the 
-    best fit for the ladder model. It computes the second derivative of the spline for each 
-    combination of peaks and selects the combination with the smallest maximum absolute value 
+    This function evaluates different combinations of size standard peaks to determine the
+    best fit for the ladder model. It computes the second derivative of the spline for each
+    combination of peaks and selects the combination with the smallest maximum absolute value
     of the second derivative, indicating the smoothest fit.
 
     Parameters:
@@ -468,10 +470,10 @@ def calculate_best_combination_of_size_standard_peaks(fsa):
     Returns:
     --------
     FsaFile
-        The FsaFile instance with the `best_size_standard` attribute updated to include 
+        The FsaFile instance with the `best_size_standard` attribute updated to include
         the best combination of size standard peaks.
     """
-    
+
     combinations = fsa.best_size_standard_combinations
 
     best_combinations = (
@@ -494,9 +496,9 @@ def fit_size_standard_to_ladder(fsa):
     """
     Fits the size standard peaks to the ladder model and updates the FSA file object.
 
-    This function iteratively fits a spline model to the best combination of size 
-    standard peaks until each base pair in the sample data is unique. The model is 
-    then used to predict the base pairs for the sample data, and the results are stored 
+    This function iteratively fits a spline model to the best combination of size
[…]
