### 2024-06-24 #6869 @willros
Hi!

Just checking in to see how everything is going? Let me know it something is unclear or anything else!

Regards,
William
