### 2024-07-04 Clinical-Genomics-Umea/fraggler#15 @KatyBrown
JOSS review - demo jupyter notebook
Issue raised as part of the JOSS review process to address the "Functionality"  check https://github.com/openjournals/joss-reviews/issues/6869

In the demo Jupyter notebook, some of the commands do not work with the version of the code currently on pip - I think the final three functions in `fraggler.py` - `fit_ladder`, `find_peaks` and `fit_area` must be elsewhere in that version as there are ImportErrors for these functions.

I therefore cloned the git repo directly and imported fraggler.fraggler from the cloned folder but this led to different errors - I would guess that the `fraggler.FsaFile` function has some different parameters in the current version of the code on GitHub than in the tutorial, as it expected `sample_channel`, `min_distance_between_peaks` and `min_size_standard_height` to be specified but they are not in the notebook, and it didn't recognise `trace_channel`.

I didn't progress past that point but it seems like the tutorial for the Python version of the code needs updating to be usable.
