### 2024-07-04 Clinical-Genomics-Umea/fraggler#16 @KatyBrown
JOSS review - API documentation
The link to the API documentation in the README.md leads to a page which I think has missing information - https://clinical-genomics-umea.github.io/fraggler/fraggler.html - here I just see the below:
![image](https://github.com/Clinical-Genomics-Umea/fraggler/assets/8896865/6d349a0a-6251-4edf-a859-cc1d27f36325)

The "fraggler" submodule link leads to the detailed code documentation describing the functions - but there is no overview of how to use the API version of the tool.
