### 2024-07-04 #6869 @KatyBrown
Hi @wilros and @csoneson, I have done a first pass of the checklist and raised a few issues - [#14](https://github.com/Clinical-Genomics-Umea/fraggler/issues/14),  [#15](https://github.com/Clinical-Genomics-Umea/fraggler/issues/15), [#16](https://github.com/Clinical-Genomics-Umea/fraggler/issues/16).

Once [#16](https://github.com/Clinical-Genomics-Umea/fraggler/issues/16) is resolved I'll be able to more thoroughly review the documentation, as currently I think I'm missing a good part of it.

Thanks :)
