### 2024-07-09 Clinical-Genomics-Umea/fraggler#15 @kdm9
Additionally, the link to `tutorial.ipynb` in the paper is broken (points to root of repo, not to the demo dir).
