### 2024-07-15 e2c2f23
added scripts and data to generate figure2
