+++ paper/figures/fraggler_manuscript_plots.Rmd
@@ -1,0 +1,161 @@
+---
+output: html_document
+editor_options: 
+  chunk_output_type: console
+---
+```{r}
+library(tidyverse)
+library(broom)
+library(cowplot)
+```
+
+# data from fraggler
+```{r}
+fraggler <- read_csv("data/fraggler_peaks.csv") %>%
+  mutate(assay_number = parse_number(assay)) %>%
+  # the quotient for assay 1 is inversed from the other assays
+  mutate(quotient = if_else(assay_number == 1 & quotient > 0, 1 / quotient, quotient))
+
+# known cnv and code for the controls 
+controls <- tibble(
+  name = c("NA18956", "NA18555", "NA12752", "NA18517", "NA10855", "NA12044", "NA07056", "NA18507"),
+  code = c(7, 6, 3, 5, 8, 2, 1, 4),
+  cnv1 = c(2, 3, 3, 4, 4, 4, 4, 5),
+  cnv2 = c(2, 2, 5, 2, 4, 5, 6, 3)
+) 
+
+fraggler <- fraggler %>%
+  # filter out the controls
+  filter(str_detect(file_name, "_K\\d_")) %>%
+  mutate(code = str_extract(file_name, "(Multi_K)(\\d*)", group = 2)) %>%
+  mutate(code = as.numeric(code)) %>%
+  group_by(file_name, code, assay_number) %>%
+  summarise(quotient = mean(quotient)) %>%
+  ungroup() %>%
+  inner_join(controls, by = "code") %>%
+  select(-file_name) %>%
+  mutate(platform = "fraggler") %>%
+  rename(ratio = quotient)
+
+```
+
+# data from peakscanner
+The peakscanner dataframe was created by uploading the files from data_for_manuscript to peakscanner online and merging the resulting csv into one file.
+```{r}
+ps <- read_csv("data/peakscanner_cnv.csv") %>%
+  janitor::clean_names() %>%
+  mutate(assay_number = parse_number(sample_filename)) %>%
+  mutate(size = as.double(size)) %>%
+  rename(area = area_base_pairs) %>%
+  select(assay_number, size, area)
+```
+
+```{r}
+quotient <- function(df, number, start, stop, n) {
+  areas <- df %>%
+    filter(assay_number == number) %>%
+    filter(between(size, start, stop)) %>%
+    slice_max(area, n = n) %>%
+    arrange(size) %>%
+    pull(area)
+    
+  areas[1] / areas[2]
+}
+
+ps_prt1 <- tibble()
+
+for (n in 1:7) {
+  ratio <- quotient(ps, n, 140, 154, 2)
+  subset <- tibble(assay_number = n, ratio = ratio, prt = 1)
+  
+  if (n == 7) {
+    subset <- tibble(assay_number = n, ratio = 0, prt = 1)
+  }
+  
+  ps_prt1 <- ps_prt1 %>% bind_rows(subset)
+} 
+
+quotient <- function(df, number, start, stop, n) {
+  areas <- df %>%
+    filter(assay_number == number) %>%
+    filter(between(size, start, stop)) %>%
+    slice_max(area, n = n) %>%
+    arrange(size) %>%
+    pull(area)
+    
+  areas[2] / areas[1]
+}
+
+ps_prt2 <- tibble()
+
+for (n in 1:7) {
+  ratio <- quotient(ps, n, 190, 200, 2)
+  subset <- tibble(assay_number = n, ratio = ratio, prt = 2)
+  
+  ps_prt2 <- ps_prt2 %>% bind_rows(subset)
+} 
+
+ps_prt3 <- tibble()
+
[…]
