### 2024-07-15 Clinical-Genomics-Umea/fraggler#14 @willros
Hi @KatyBrown!

Thanks for the input. I have now added files and scripts inside the `paper/figures` directory which can be used to generate Figure 2.

Generating the files using the Peakscanner software demands some manual work and is quite cumbersome (one of the reason we created Fraggler), but I have described how this was done in the `FIGURES.md`.

Let me know if you want anything else regarding this issue!
William
