### 2024-07-15 Clinical-Genomics-Umea/fraggler#15 @willros
Hi @KatyBrown and @kdm9!

I have updated the version of fraggler on pypi. The tutorial works for me when I pip install the latest version of fraggler.
Try it out and let me know if it still does not work.

William
