### 2024-07-15 Clinical-Genomics-Umea/fraggler#16 @willros
Hi @KatyBrown!

The API of Fraggler is best demonstrated in the `tutorial.ipynb`. This notebook documents everything needed to use Fraggler in a Python environment.

The link to the documentation provides a comprehensive overview of the source code. I have tried to document and describe the code thoroughly, making it easy to understand the logic and flow.

Hope this helps!

William
