### 2024-08-06 Clinical-Genomics-Umea/fraggler#15 @KatyBrown
Thank you, the tutorial works nicely now.
