### 2024-08-06 Clinical-Genomics-Umea/fraggler#16 @KatyBrown
I see,  I was expecting that the link above was the main documentation of the API.
