### 2024-08-21 #6869 @kdm9
Hi @willros, this time around I was able to completely run the tutorial from a fresh debian 12 docker container, so I can confirm all is working great. I'm happy to recommend accepting.

A couple of very minor notes/nitpicks:

- API docs: they exist, but because of your module structure take a bit of clicking around to get to. I'd suggest linking directly to <https://clinical-genomics-umea.github.io/fraggler/fraggler/fraggler.html> in the readme, or adding a 'landing page' docstring to the base fraggler module to explain that the code is predominantly in fraggler.fraggler. You can probably also omit some of the simplest functions like `print_green` from the api docs.
- Your CLI prints some nice graphics and coloured texts. Personally, I prefer to only print these when [stderr/stdout is a tty](https://stackoverflow.com/questions/1077113/how-do-i-detect-whether-sys-stdout-is-attached-to-terminal-or-not/1077118#1077118), but not for log files. Viewing the log with less will show the non-printing terminal control characters, which makes the log hard to read

Apologies again for the delay!

best,
K
