### 2024-08-21 #6869 @willros
Hi @kdm9!

Thanks for the review! Thanks for the suggested improvements. I linked the API docs to the page you suggested. I also added a decorator for printing with colors if terminal if tty. If not, the text is just printed in a regular fashion. I agree, this makes the log much easier to read! Thanks for pointing that out.

Regards,
William
