### 2024-08-21 #6869 @kdm9
[One last comment](https://github.com/Clinical-Genomics-Umea/fraggler/commit/4d330f2623e31e28923fe856383dddaeaabab99e#r145614505) on that commit, but otherwise all perfect from my end!
