### 2024-08-24 813b6d0
Spell out journal names
