+++ paper/paper.bib
@@ -31,7 +31,7 @@
               products derived from the CNV region versus the external
               reference. In contrast to other approaches, setting up this
               genotyping method requires an affordable investment.",
-  journal  = "Mol. Cell. Probes",
+  journal  = "Molecular and Cellular Probes",
   volume   =  29,
   number   =  6,
   pages    = "517--521",
@@ -49,7 +49,7 @@
   title    = "Accurate, high-throughput typing of copy number variation using
               paralogue ratios from dispersed repeats",
   author   = "Armour, John A L and Palla, Raquel and Zeeuwen, Patrick L J M and
-              den Heijer, Martin and Schalkwijk, Joost and Hollox, Edward J",
+              {den Heijer}, Martin and Schalkwijk, Joost and Hollox, Edward J",
   abstract = "Recent work has demonstrated an unexpected prevalence of copy
               number variation in the human genome, and has highlighted the
               part this variation may play in predisposition to common
@@ -72,7 +72,7 @@
               method for copy number analysis, suitable for application to
               typing thousands of samples in large case-control association
               studies.",
-  journal  = "Nucleic Acids Res.",
+  journal  = "Nucleic Acids Research",
   volume   =  35,
   number   =  3,
   pages    = "e19",
@@ -125,7 +125,7 @@
               popular software for fragment analysis while possessing unique
               advantages and the possibility of automation for high-throughput
               experiments by exploiting the power of R.",
-  journal  = "BMC Genet.",
+  journal  = "BMC Genetics",
   volume   =  17,
   pages    = "62",
   month    =  apr,
@@ -164,7 +164,7 @@
               and investigation of its relationship with malaria, in large
               sample sets at minimal cost, using standard molecular biology
               equipment.",
-  journal  = "Ann. Hum. Genet.",
+  journal  = "Annals of Human Genetics",
   volume   =  85,
   number   =  1,
   pages    = "7--17",
@@ -181,7 +181,7 @@
 @ARTICLE{Van_Steenderen2022-dz,
   title    = "{BinMat}: A molecular genetics tool for processing binary data
               obtained from fragment analysis in {R}",
-  author   = "van Steenderen, Clarke",
+  author   = "{van Steenderen}, Clarke",
   abstract = "Processing and visualising trends in the binary data (presence or
               absence of electropherogram peaks), obtained from fragment
               analysis methods in molecular biology, can be a time-consuming
@@ -199,7 +199,7 @@
               to visualise their data as ordination plots and clustering trees
               without having to use multiple programmes and input files or rely
               on previous programming experience.",
-  journal  = "Biodivers Data J",
+  journal  = "Biodiversity Data Journal",
   volume   =  10,
   pages    = "e77875",
   month    =  mar,
@@ -237,7 +237,7 @@
               multiple formats and can serve as a pre-processing suite that
               prepares data for statistical analysis in programs such as SAS or
               R.",
-  journal  = "Environ. Microbiol. Rep.",
+  journal  = "Environmental Microbiology Reports",
   volume   =  4,
   number   =  5,
   pages    = "556--561",
@@ -374,7 +374,7 @@
               repositories and millions of downloads per year. In this work, we
               provide an overview of the capabilities and development practices
               of SciPy 1.0 and highlight some recent technical developments.",
-  journal  = "Nat. Methods",
+  journal  = "Nature Methods",
   volume   =  17,
   number   =  3,
   pages    = "261--272",
@@ -419,7 +419,7 @@
               shaped the pattern of CNV at DMBT1, and that the DMBT1-S. mutans
               interaction is a promising model of host-pathogen-culture
               coevolution in humans.",
[…]
