+++ paper/paper.bib
@@ -250,107 +250,6 @@
   doi      = "10.1111/j.1758-2229.2012.00368.x"
 }
 
-@ARTICLE{Walker2009-of,
-  title    = "Multiplex Paralogue Ratio Tests for accurate measurement of
-              multiallelic {CNVs}",
-  author   = "Walker, Susan and Janyakhantikul, Somwang and Armour, John A L",
-  abstract = "Demonstrating an association between a polymorphism and a disease
-              phenotype through case-control studies requires reliable
-              large-scale genotyping, but accurate measurement of copy number
-              variation has proven to be technically challenging. Here we build
-              on our previous experience with Paralogue Ratio Tests (PRT) to
-              develop PRT copy number determination at the CCL3L1/CCL4L1 copy
-              number variant. A multiplex PRT assay based on four independent
-              comparative PCRs results in a convenient, accurate and robust
-              method of multiallelic copy number measurement suitable for use
-              in large-scale case-control studies, which can unambiguously
-              assign virtually all samples tested to discrete copy number
-              classes.",
-  journal  = "Genomics",
-  volume   =  93,
-  number   =  1,
-  pages    = "98--103",
-  month    =  jan,
-  year     =  2009,
-  keywords = "Fragment analysis",
-  language = "en",
-  issn     = "0888-7543, 1089-8646",
-  pmid     = "18848619",
-  doi      = "10.1016/j.ygeno.2008.09.004"
-}
-
-@ARTICLE{Cock2009-ai,
-  title    = "Biopython: freely available Python tools for computational
-              molecular biology and bioinformatics",
-  author   = "Cock, Peter J A and Antao, Tiago and Chang, Jeffrey T and
-              Chapman, Brad A and Cox, Cymon J and Dalke, Andrew and Friedberg,
-              Iddo and Hamelryck, Thomas and Kauff, Frank and Wilczynski,
-              Bartek and de Hoon, Michiel J L",
-  abstract = "SUMMARY: The Biopython project is a mature open source
-              international collaboration of volunteer developers, providing
-              Python libraries for a wide range of bioinformatics problems.
-              Biopython includes modules for reading and writing different
-              sequence file formats and multiple sequence alignments, dealing
-              with 3D macro molecular structures, interacting with common tools
-              such as BLAST, ClustalW and EMBOSS, accessing key online
-              databases, as well as providing numerical methods for statistical
-              learning. AVAILABILITY: Biopython is freely available, with
-              documentation and source code at (www.biopython.org) under the
-              Biopython license.",
-  journal  = "Bioinformatics",
-  volume   =  25,
-  number   =  11,
-  pages    = "1422--1423",
-  month    =  jun,
-  year     =  2009,
-  language = "en",
-  issn     = "1367-4803, 1367-4811",
-  pmid     = "19304878",
-  doi      = "10.1093/bioinformatics/btp163",
-  pmc      = "PMC2682512"
-}
-
-@ARTICLE{Sessions2015-ml,
-  title    = "Analysis of Dengue Virus Genetic Diversity during Human and
-              Mosquito Infection Reveals Genetic Constraints",
-  author   = "Sessions, October M and Wilm, Andreas and Kamaraj, Uma Sangumathi
-              and Choy, Milly M and Chow, Angelia and Chong, Yuwen and Ong, Xin
-              Mei and Nagarajan, Niranjan and Cook, Alex R and Ooi, Eng Eong",
-  abstract = "Dengue viruses (DENV) cause debilitating and potentially
-              life-threatening acute disease throughout the tropical world.
-              While drug development efforts are underway, there are concerns
-              that resistant strains will emerge rapidly. Indeed, antiviral
-              drugs that target even conserved regions in other RNA viruses
-              lose efficacy over time as the virus mutates. Here, we sought to
[…]
