### 2024-08-24 #6869 @csoneson
👋🏻 @willros - I have gone through your submission; overall it looks good, I have made a small [PR](https://github.com/Clinical-Genomics-Umea/fraggler/pull/21) with some suggested edits. Please check if you agree (and that I didn't inadvertently change the intended meaning of the text) and if so, merge and regenerate the proof here using `@editorialbot generate pdf`. I will also make another post below with the next steps.
