### 2024-08-25 #6869 @willros
@csoneson

Thanks for the suggested changes to the text. I merged your PR.

* The latest version is v.3.0.3

* Here is the archived repository: [10.5281/zenodo.13370670](https://zenodo.org/doi/10.5281/zenodo.13370670)
