### 2024-08-25 #6869 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
