### 2024-08-26 #6869 @Kevin-Mattheus-Moerman
@willros as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I have checked this review, your repository, the archive link, and the paper. Everything seems to be in order, so I will now proceed to accept this submission. Well done everybody.
