### 2024-08-26 #6869 @willros
@kdm9 @csoneson @KatyBrown
Thanks everyone! It has been a very nice process to publish with joss!

Regards,
William
